4755 доработки по задаче

+++ Source/Libraries/Core/Backend/CustomerOrdersApi.Library/Dto/Orders/OnlineRentPackageDto.cs
@@ -18,6 +18,6 @@
 		/// <summary>
 		/// Количество
 		/// </summary>
-		public decimal Count { get; set; }
+		public int Count { get; set; }
 	}
 }
